refactor pipeline in jedis-plugin
